Refactor layout
- Move core app functionality to bmc
- Split routes to separate file
- Move template directory to bmc
