+++ bmc/__init__.py
@@ -1,0 +1,5 @@
+from flask import Flask
+
+bmc_app = Flask(__name__)
+
+from bmc import routes
